Add initial README.md
This commit adds a README.md file to the repository, providing an overview of my skills, current role, and how to contact me. The README includes information on my work as a Pro Hacker at Hack The Box, my web development expertise, and my interest in collaborating on innovative tech projects.